pyproject.toml
[project]
name = "library-api"
version = "0.1.0"
description = "API to store my book reviews"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "pydantic-settings>=2.7.1",
    "sqlalchemy[asyncio]>=2.0.37",
    "alembic>=1.14.1",
    "gunicorn>=23.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest-asyncio>=0.25.2",
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
    "nest-asyncio>=1.6.0",
    "testcontainers[postgres]>=4.9.0",
    "ruff>=0.9.3",
    "factory-boy>=3.3.3",
    "faker>=37.1.0",
    "isort>=6.0.1",
    "async-factory-boy>=1.0.1",
    "pre-commit>=4.2.0",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"

[tool.isort]
profile = "black"
line_length = 80
extend_skip_glob = ["infrastructure/*"]

[tool.ruff]
line-length = 80
extend-exclude = ["infrastructure"]
src = ["app", "tests"]

[tool.ruff.lint]
extend-select = ["F"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
